\name{RcppTomlPlusPlus-package}
\alias{RcppTomlPlusPlus-package}
\alias{RcppTomlPlusPlus}
\docType{package}
\title{\packageTitle{RcppTomlPlusPlus}}
\description{\packageDescription{RcppTomlPlusPlus}}
\section{Package Content}{\packageIndices{RcppTomlPlusPlus}}
\author{\packageAuthor{RcppTomlPlusPlus}}
\section{Maintainer}{\packageMaintainer{RcppTomlPlusPlus}}
\keyword{package}
