% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simpleParser}
\alias{simpleParser}
\title{Simple Example Parser}
\usage{
simpleParser(filename)
}
\arguments{
\item{filename}{An existing TOML file}
}
\value{
An integer indicating suggess or failure
}
\description{
This example is adapted from the upstream documentation and simply prints
an existing TOML file to standard output.
}
\examples{
file <- system.file("toml", "example.toml", package="RcppTomlPlusPlus")
simpleParser(file)
}
