% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tomlparseImpl}
\alias{tomlparseImpl}
\title{Parse a TOML file}
\usage{
tomlparseImpl(input, verbose = FALSE, fromfile = TRUE, includize = FALSE,
  escape = TRUE)
}
\arguments{
\item{input}{[character] TOML input, either as chracter value or path to TOML file}

\item{verbose}{[logical] Optional verbosity flag, no current effect}

\item{fromfile}{[logical] Optional with default value \sQuote{TRUE} indicating parsing from file}

\item{includize}{[logical] Optional legacy option, no current effect}

\item{escape}{[logical] Optiona legacy option, no current effect}
}
\value{
A List with the parsed content
}
\description{
Parse a TOML file
}
